+++ packages/x-charts/src/RadarChart/RadarAxis/radarAxisClasses.ts
@@ -1,3 +1,4 @@
+import composeClasses from '@mui/utils/composeClasses';
 import generateUtilityClass from '@mui/utils/generateUtilityClass';
 import generateUtilityClasses from '@mui/utils/generateUtilityClasses';
 
@@ -20,3 +21,13 @@ export const chartsAxisClasses: RadarAxisClasses = generateUtilityClasses('MuiRa
   'line',
   'label',
 ]);
+
+export const useUtilityClasses = (classes?: Partial<RadarAxisClasses>) => {
+  const slots = {
+    root: ['root'],
+    line: ['line'],
+    label: ['label'],
+  };
+
+  return composeClasses(slots, getRadarAxisUtilityClass, classes);
+};
