+++ scripts/x-charts-pro.exports.json
@@ -264,7 +264,6 @@
   { "name": "getPieArcLabelUtilityClass", "kind": "Function" },
   { "name": "getPieArcUtilityClass", "kind": "Function" },
   { "name": "getPieCoordinates", "kind": "Function" },
-  { "name": "getRadarAxisUtilityClass", "kind": "Function" },
   { "name": "getReferenceLineUtilityClass", "kind": "Function" },
   { "name": "getValueToPositionMapper", "kind": "Function" },
   { "name": "greenPalette", "kind": "Variable" },
@@ -398,7 +397,6 @@
   { "name": "purplePaletteLight", "kind": "Variable" },
   { "name": "RadarAxis", "kind": "Function" },
   { "name": "RadarAxisClasses", "kind": "Interface" },
-  { "name": "RadarAxisClassKey", "kind": "TypeAlias" },
   { "name": "RadarAxisHighlight", "kind": "Function" },
   { "name": "RadarAxisHighlightProps", "kind": "Interface" },
   { "name": "RadarAxisProps", "kind": "Interface" },
